requirements.txt
aniso8601==9.0.1
attrs==22.1.0
click==8.1.3
Flask==2.1.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
flask-restx==0.5.1
Flask-Security-Too==5.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3.post0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.4
pyrsistent==0.18.1
python-dotenv==0.21.0
pytz==2022.4
six==1.16.0
SQLAlchemy==1.4.41
Werkzeug==2.1.2
